third_party/WebKit/LayoutTests/media/video-enter-fullscreen-without-user-gesture.html

<!DOCTYPE html>
<title>Test webkitRequestFullscreen() without any user gesture.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.oncanplaythrough = t.step_func(function() {
        // No user gesture initiated.
        assert_true(video.paused);
        video.webkitRequestFullscreen();
        video.onwebkitfullscreenerror = t.step_func_done();
    });

    video.src = "content/test.ogv";
});
</script>
